+++ includes/Math/Math.hpp
@@ -6,7 +6,7 @@
 /*   By: mbatty <mbatty@student.42angouleme.fr>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/01/02 13:00:48 by mbatty            #+#    #+#             */
-/*   Updated: 2026/01/02 13:13:47 by mbatty           ###   ########.fr       */
+/*   Updated: 2026/01/02 15:17:46 by mbatty           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,12 +14,13 @@
 
 #include "Vec3.hpp"
 #include "Vec2.hpp"
+#include "Mat4.hpp"
 #include <cmath>
 
 // Vec3
 
-double	dot(const Vec3 &v1, const Vec3 &v2);
-double	length(const Vec3 &v);
+float	dot(const Vec3 &v1, const Vec3 &v2);
+float	length(const Vec3 &v);
 Vec3	cross(const Vec3 &v1, const Vec3 &v2);
 Vec3	normalize(const Vec3 &v);
