+++ includes/Math/Vec2.hpp
@@ -6,7 +6,7 @@
 /*   By: mbatty <mbatty@student.42angouleme.fr>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/01/02 12:59:53 by mbatty            #+#    #+#             */
-/*   Updated: 2026/01/02 13:13:41 by mbatty           ###   ########.fr       */
+/*   Updated: 2026/01/02 15:17:43 by mbatty           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,15 +19,15 @@ struct	Vec2
 		this->x = 0;
 		this->y = 0;
 	}
-	Vec2(double x, double y)
+	Vec2(float x, float y)
 	{
 		this->x = x;
 		this->y = y;
 	}
-	Vec2(double xy)
+	Vec2(float xy)
 	{
 		this->x = xy;
 		this->y = xy;
 	}
-	double	x, y;
+	float	x, y;
 };
